pyproject.toml
[project]
name = "AutoOrgancommercial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "coloredlogs==15.0.1",
    "cryptography==46.0.3",
    "cucim-cu12>=25.10.0",
    "cupy-cuda12x>=13.6.0",
    "ecdsa==0.19.1",
    "fastapi==0.121.1",
    "fastrlock==0.8.3",
    "filelock==3.19.1",
    "flatbuffers==25.9.23",
    "fsspec==2025.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "humanfriendly==10.0",
    "idna==3.11",
    "imageio==2.37.2",
    "lazy-loader==0.4",
    "loguru==0.7.3",
    "markupsafe==2.1.5",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "networkx==3.5",
    "nibabel==5.3.2",
    "numpy==1.26.4",
    "onnxruntime-gpu==1.20.2",
    "packaging==25.0",
    "pillow==11.3.0",
    "protobuf==6.33.0",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pyjwt>=2.10.1",
    "python-jose==3.5.0",
    "requests==2.32.5",
    "rsa==4.9.1",
    "scikit-image==0.25.2",
    "scipy==1.16.3",
    "setuptools==80.9.0",
    "simpleitk==2.5.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.49.3",
    "sympy==1.14.0",
    "tifffile==2025.10.16",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uv==0.9.8",
    "uvicorn==0.38.0",
    "wheel==0.45.1",
]

